+++ __init__.py
@@ -1,5 +1,0 @@
-"""Initialize PyPI Package"""
-
-# pylint: disable=locally-disabled, invalid-name, import-error
-
-__all__ = []