+++ __main__.py
@@ -1,8 +1,0 @@
-"""Run widgets from main."""
-
-# pylint: disable=locally-disabled, invalid-name, import-error
-
-from main import widgets
-
-if __name__ == "__main__":
-    widgets()